Pre: move EmptyBecause into types.rs.
